bump swagger version. support custom swagger config. regiter crud operations differently.

+++ ob1k-swagger/src/main/java/com/outbrain/ob1k/swagger/service/SwaggerService.java
@@ -2,6 +2,7 @@ package com.outbrain.ob1k.swagger.service;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.google.common.collect.Sets;
 import com.outbrain.ob1k.HttpRequestMethodType;
 import com.outbrain.ob1k.Request;
 import com.outbrain.ob1k.Response;
@@ -13,12 +14,19 @@ import com.outbrain.ob1k.server.registry.ServiceRegistryView;
 import com.outbrain.ob1k.server.registry.endpoints.ServerEndpointView;
 import io.netty.handler.codec.http.HttpResponseStatus;
 import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiImplicitParam;
+import io.swagger.annotations.ApiImplicitParams;
 import io.swagger.annotations.ApiParam;
+import io.swagger.converter.ModelConverters;
 import io.swagger.models.Info;
+import io.swagger.models.Model;
 import io.swagger.models.Operation;
 import io.swagger.models.Path;
 import io.swagger.models.Swagger;
 import io.swagger.models.Tag;
+import io.swagger.models.parameters.BodyParameter;
+import io.swagger.models.parameters.HeaderParameter;
+import io.swagger.models.parameters.PathParameter;
 import io.swagger.models.parameters.QueryParameter;
 
 import java.io.IOException;
@@ -26,6 +34,7 @@ import java.io.StringWriter;
 import java.lang.annotation.Annotation;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 
 import static com.google.common.collect.Lists.asList;
 import static io.netty.handler.codec.http.HttpHeaders.Names.CONTENT_TYPE;
@@ -53,7 +62,10 @@ public class SwaggerService implements Service {
   private Swagger buildSwagger(final Request request) {
     final Swagger swagger = new Swagger();
     swagger.host(request.getHeader("Host"));
-    swagger.info(buildInfo());
+    swagger.info(buildInfo());                          
+
+
+    Set<ISwaggerAware> invoked = Sets.newHashSet();
     for (final Map.Entry<String, Map<HttpRequestMethodType, ServerEndpointView>> entry :
             serviceRegistry.getRegisteredEndpoints().entrySet()) {
       final Path path = new Path();
@@ -61,27 +73,35 @@ public class SwaggerService implements Service {
         final HttpRequestMethodType methodType = endpointEntry.getKey();
         final String key = entry.getKey();
         final ServerEndpointView endpoint = endpointEntry.getValue();
+        final Service service = endpoint.service();
         if (!ignoreEndpoint(endpoint)) {
           final Tag tag = buildTag(endpoint.getMethod().getDeclaringClass());
           swagger.addTag(tag);
-          switch (methodType) {
-            case GET:
-            case ANY:
-              path.get(buildOperation(endpoint, tag, methodType));
-              break;
-            case POST:
-              path.post(buildOperation(endpoint, tag, methodType));
-              break;
-            case PUT:
-              path.put(buildOperation(endpoint, tag, methodType));
-              break;
-            case DELETE:
-              path.delete(buildOperation(endpoint, tag, methodType));
-              break;
-            default:
-              throw new UnsupportedOperationException("Unsupported method type " + methodType);
+          if (service instanceof ISwaggerAware) {
+            ISwaggerAware swaggerAware = (ISwaggerAware) service;
+            if(invoked.add(swaggerAware)) {
+              swaggerAware.invoke(swagger, key);
+            }
+          } else {
+            switch (methodType) {
+              case GET:
+              case ANY:
+                path.get(buildOperation(endpoint, tag, methodType));
+                break;
+              case POST:
+                path.post(buildOperation(endpoint, tag, methodType));
+                break;
+              case PUT:
+                path.put(buildOperation(endpoint, tag, methodType));
+                break;
+              case DELETE:
+                path.delete(buildOperation(endpoint, tag, methodType));
+                break;
+              default:
+                throw new UnsupportedOperationException("Unsupported method type " + methodType);
+            }
+            swagger.path(key, path);
           }
-          swagger.path(key, path);
         }
       }
     }
@@ -108,11 +128,52 @@ public class SwaggerService implements Service {
     final Operation operation = new Operation().summary(endpoint.getTargetAsString()).tag(tag.getName()).
             operationId(endpoint.getTargetAsString() + "Using" + methodType.name());
     final String[] endpointParamNames = endpoint.getParamNames();
+
+    ApiImplicitParams implicitParamAnnotation = findAnnotation(endpoint.getMethod().getAnnotations(), ApiImplicitParams.class);
+    if (implicitParamAnnotation != null) {
+      for (ApiImplicitParam param : implicitParamAnnotation.value()) {
+        String dataType = param.dataType();
+        String paramName = param.name();
+        String paramType = param.paramType();
+        if (paramType.equals("body")) {
+          Class<?> type = null;
+          try {
+            type = Class.forName(dataType);
+          } catch (ClassNotFoundException e) {
+            throw new RuntimeException(e.getMessage(), e);
+          }
+          Map<String, Model> model = ModelConverters.getInstance().read(type);
+          if (model.size() == 1) {
+            BodyParameter bodyParameter = new BodyParameter().name(paramName).schema(model.values().iterator().next());
+            bodyParameter.setRequired(true);
+            operation.addParameter(bodyParameter);
+          }
+
+        } else if (paramType.equals("path")) {
+          PathParameter pathParameter = new PathParameter().name(paramName).type(dataType);
+          pathParameter.setRequired(true);
+          operation.addParameter(pathParameter);
+        } else if (paramType.equals("header")) {
+          HeaderParameter pathParameter = new HeaderParameter().name(paramName).type(dataType);
+          pathParameter.setRequired(param.required());
+          operation.addParameter(pathParameter);
+        } else if (paramType.equals("query")) {
+          QueryParameter pathParameter = new QueryParameter().name(paramName).type(dataType);
+          pathParameter.setRequired(param.required());
+          operation.addParameter(pathParameter);
+        }
+      }
+    }
+
     final Annotation[][] parameterAnnotations = endpoint.getMethod().getParameterAnnotations();
     final Class<?>[] parameterTypes = endpoint.getMethod().getParameterTypes();
     for (int i = 0; i < endpointParamNames.length; i++) {
-      final ApiParam annotation = findApiParamAnnotation(parameterAnnotations[i]);
-      final String type = getSwaggerDataType(parameterTypes[i]);
+      Class<?> parameterType = parameterTypes[i];
+      if (Request.class.equals(parameterType) && implicitParamAnnotation != null) {
+        continue;
+      }
+      final ApiParam annotation = findAnnotation(parameterAnnotations[i], ApiParam.class);
+      final String type = getSwaggerDataType(parameterType);
       final String paramName = (annotation != null) ? annotation.name() : endpointParamNames[i];
       final QueryParameter param = new QueryParameter().type(type).name(paramName);
       if (annotation != null) {
@@ -123,10 +184,10 @@ public class SwaggerService implements Service {
     return operation;
   }
 
-  private ApiParam findApiParamAnnotation(Annotation[] annotations) {
+  private <T> T findAnnotation(Annotation[] annotations, Class<T> t) {
     for (Annotation annotation : annotations) {
-      if (ApiParam.class.equals(annotation.annotationType())) {
-        return (ApiParam) annotation;
+      if (t.equals(annotation.annotationType())) {
+        return t.cast(annotation);
       }
     }
     return null;
